Network.MQTT.RPC

Documentation

call :: MonadIO m => MQTTClient -> Topic -> ByteString -> m ByteString

Send a message to a topic on an MQTT broker with a random subscription and correlation such that an agent may receive this message and respond over the ephemeral channel. The response will be returned.

Note that this client provides no timeouts or retries. MQTT will guarantee the request message is delivered to the broker, but if there's nothing to pick it up, there may never be a response.